Convert command preload to oclif, typescript
Change-type: patch
