+++ package.json
@@ -181,6 +181,7 @@
     "nock": "^12.0.3",
     "parse-link-header": "~1.0.1",
     "pkg": "^4.4.9",
+    "pinejs-client-core": "^6.7.3",
     "publish-release": "^1.6.1",
     "rewire": "^4.0.1",
     "simple-git": "^1.132.0",
